backend/requirements.txt
Flask==1.0.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
grpcio-tools==1.19.0
Authlib==0.10
loginpass==0.2.1
python-dotenv==0.10.1
gunicorn>=19.1.1
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.10
Flask-Migrate==2.2.1
psycopg2==2.7.5
marshmallow==3.0.0b13
flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.16.1
base58==0.2.5
secp256k1==0.13.2
bitstring==3.1.5
webargs==5.2.0
flask-cors==3.0.7
